Moved rule tests to `rules` directory. Fixed code alignment in tests. Fixed typo.
